+++ runtime/moonbeam/src/xcm_config.rs
@@ -691,7 +691,7 @@ pub type ForeignAssetManagerOrigin = EitherOfDiverse<
 parameter_types! {
 	/// Balance in the native currency that will be reserved from the user
 	/// to create a new foreign asset
-	pub ForeignAssetCreationDeposit: u64 =
+	pub ForeignAssetCreationDeposit: u128 =
 		runtime_params::dynamic_params::xcm_config::ForeignAssetCreationDeposit::get();
 }
